A thread now goes to sleep if receives task=nullptr
